+++ packages/subgraph/CHANGELOG.md
@@ -7,12 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UPCOMING]
 
+### Added
+- Adds `startDate` and `endDate` fields to Multisig proposals
+
 ## 0.6.1-alpha
 On 2023-01-11 17:06:50
 
 ### Changed
 - Ignores `None` votes from addresslist voting and token voting
 - Updates `MULTISIG_INTERFACE`
+
+### Removed
+- Removes `open` and `executable` fields from Multisig proposals 
 
 ## 0.5.0-alpha